Fix Python Xprt object reference
The bug was revealed by `ldms/python/test/ldmsd/ldmsd_failover.py`. The
Python Xprt object is a local variable and is used in a blocking mode in
the following sequence: connect, dir, discard (leave it be). The problem
is that the DISCONNECTED event could then be delivered from zap thread
after the python thread has already finished the function that uses that
Python Xprt local variable. Hence, the Python Xprt object has already
gone by the time DISCONNECTED is delivered ... causing a segmentation
fault.

This patch fixes the bug by taking a reference when the transport
becomes CONNECTED, and put it back when DISCONNECTED is delivered. We
don't have to worry about other events in the blocking mode because
DISCONNECTED is guaranteed to be delivered by LDMS only if CONNECTED has
been delivered. For the non-blocking mode (application supplied
callback), we let the application takes care of the Python Xprt
reference.